+++ src/audio/codec/mod.rs
@@ -8,7 +8,12 @@ mod wm8731;
 #[cfg(feature = "seed_1_1")]
 pub use wm8731::{Codec, Pins};
 
+#[cfg(feature = "seed_1_2")]
+mod pcm3060_parallel;
+#[cfg(feature = "seed_1_2")]
+pub use pcm3060_parallel::{Codec, Pins};
+
 #[cfg(feature = "patch_sm")]
-mod pcm3060;
+mod pcm3060_i2c;
 #[cfg(feature = "patch_sm")]
-pub use pcm3060::{Codec, Pins};
+pub use pcm3060_i2c::{Codec, Pins};